Final Changes
"Mob Stomping" has been implimented.  Stomping a mob now results in the mobs death.  Any other contact with mobs results in the player's death.

Some parts of the contact listener were cleaned up to remove unnecessary/unused lines.

Whether the last level has ended (and thus the game) is now checked based on the size of the levels vector instead of a constant in definitions (thus allowing the game to be more dynamic).

Fast fall speed was reduced to a more resonable level.  This also helps avoid the player clipping to far into a mob while stomping before the contact listener is checked resulting in the mob stomp check failing and the player dying.

Throwing sound effect is now deleted from the library when it is destroyed.

Throwing sound effect now plays on the first available channel to allow for more than one instance to play in the event that there is more than one crazy person nearby.

Unneeded FIXMEs removed.

Various sections of code had their comments adjusted to be clearer.